refactor: extract session runtime modules
